Sorted out the error handling (for now).
On one hand, it works just fine for what I want it for.

On the other hand, I lose the functional dependency in the error classes
because I was struggling to thread things through in a way that would
meet the coverage conditions.

I'll return to this at some point and try to have my cake and eat it too.